Make Algo.map return an int to avoid floating-point conversion issues

+++ software/contrib/pet_rock.py
@@ -87,7 +87,7 @@ class Algo:
     def map(x, in_min, in_max, out_min, out_max):
         # treat the output as inclusive
         out_max = out_max + 1
-        return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min
+        return int((x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min)
 
     def tick(self):
         """
